fix(webgl2): use highp float for all vertex position/UV stages
Mobile GLES honours the vertex-stage precision qualifier: lowp float has
range +/-2 and collapses world coordinates, while mediump degrades them.
Desktop drivers ignore the qualifier, so this is a zero-diff change there
but a correctness fix on real mobile hardware (finding F4).

Promote every vertex stage that does position/UV math to highp float:
sprite, mesh, particle, mask-compose and backdrop-blend GLSL files, plus
the inline nine-slice and repeating (shader + geometry path) sources.
text.vert and stencil-clip.vert were already highp.

+++ src/rendering/webgl2/WebGl2RepeatingSpriteRenderer.ts
@@ -18,7 +18,7 @@ import { WebGl2VertexArrayObject, type WebGl2VertexArrayObjectRuntime } from './
 // ---------------------------------------------------------------------------
 
 const shaderPathVertSource = `#version 300 es
-precision mediump float;
+precision highp float;
 precision highp int;
 
 layout(location = 0) in vec4 a_quadBounds;   // x0,y0,x1,y1 (local space)
@@ -68,7 +68,7 @@ void main(void) {
 // ---------------------------------------------------------------------------
 
 const geoPathVertSource = `#version 300 es
-precision lowp float;
+precision highp float;
 precision highp int;
 
 layout(location = 0) in vec4 a_quadBounds;   // x0,y0,x1,y1 (local space)
